# Use vectorizable block function