This is a port of the vectorizable rust-crypto block function and its accompanying simd utility library. Pretty much just a copy and paste job, I'm not quite sure how to properly mark copyrights in cases like these. I have updated the copyright notice in the simd.rs file to be the MIT license, it is also available under apache-2.

I also updated the openssl library for tests because I couldnt get the old version to compile. If you want I can remove that update.

Benchmarks: Before:
```
test digest::sha1::_1000     ... bench:  6,046 ns/iter (+/- 2,362) = 165 MB/s
test digest::sha1::_16       ... bench:    389 ns/iter (+/- 130) = 41 MB/s
test digest::sha1::_2000     ... bench: 11,979 ns/iter (+/- 2,491) = 166 MB/s
test digest::sha1::_256      ... bench:  1,898 ns/iter (+/- 776) = 134 MB/s
test digest::sha1::_8192     ... bench: 48,230 ns/iter (+/- 19,805) = 169 MB/s
test digest::sha1::block_len ... bench:    755 ns/iter (+/- 252) = 84 MB/s
```

After:
```
test digest::sha1::_1000     ... bench:       2,239 ns/iter (+/- 307) = 446 MB/s
test digest::sha1::_16       ... bench:         156 ns/iter (+/- 15) = 102 MB/s
test digest::sha1::_2000     ... bench:       4,431 ns/iter (+/- 534) = 451 MB/s
test digest::sha1::_256      ... bench:         713 ns/iter (+/- 78) = 359 MB/s
test digest::sha1::_8192     ... bench:      17,928 ns/iter (+/- 1,749) = 456 MB/s
test digest::sha1::block_len ... bench:         294 ns/iter (+/- 33) = 217 MB/s
```