@kkartaltepe 2017-07-07
Any comments on this PR?